# workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/workspace/cpg.bin/server/api.yaml
openapi: "3.0.3"
info:
  title: Call Center Broker API
  version: "1.0.0"
  description: >
    Broker API that dispatches owner leads to agent mobile devices.
    Agents authenticate, receive owners to call, and submit call outcomes.
servers:
  - url: https://api.server.test/api/v1
  - url: http://localhost:3000/api/v1
    description: Local dev
security:
  - BearerAuth: []
  # - apiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    LoginResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
          example: "eyJhbGci..."
        user:
          $ref: "#/components/schemas/User"
...
    User:
      type: object
      required: [id, email, role]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        role:
          type: string
        has_profile_image:
          type: boolean
          description: Whether the user has a profile image stored
        is_online:
          type: boolean
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "OWNER_NOT_FOUND"
        message:
          type: string
          example: "No owner with that ID exists"
        details:
          type: object
          nullable: true
  responses:
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ValidationError:
      description: Request body or params failed validation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
paths:
  /login:
    post:
      summary: Agent / admin login
      operationId: login
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
      responses:
        "200":
          description: JWT issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /register:
    post:
      summary: Register a new user (auto-login)
      operationId: register
      tags: [auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                password:
                  type: string
                  format: password
                name:
                  type: string
                phone:
                  type: string
                role:
                  type: string
      responses:
        "201":
          description: User created and logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
  /logout:
    post:
      summary: Invalidate the current session
      operationId: logout
      tags: [auth]
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Logged out successfully"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /me:
    get:
      summary: Get the currently authenticated user
      operationId: getMe
      tags: [auth]
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"
...